修改pipetest，使其测试了async + await、async + non-await、non-async + non-await三种实现方式
